Replace older view tests with unit tests
